Renamed the class path in the accumulo-spark package to geotrellis.spark.store.accumulo
