+++ accumulo-spark/src/main/resources/META-INF/services/geotrellis.spark.io.LayerReaderProvider
@@ -1,1 +1,1 @@
-geotrellis.spark.io.accumulo.AccumuloSparkLayerProvider
+geotrellis.spark.store.accumulo.AccumuloSparkLayerProvider
