+++ accumulo-spark/src/test/scala/geotrellis/spark/store/accumulo/MockAccumuloInstance.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.accumulo
+package geotrellis.spark.store.accumulo
 
-import geotrellis.layers.accumulo._
+import geotrellis.store.accumulo._
 import org.apache.accumulo.core.client._
 import org.apache.accumulo.core.client.mock.MockInstance
 import org.apache.accumulo.core.client.security.tokens.PasswordToken
