+++ accumulo-store/src/main/resources/reference.conf
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-geotrellis.store.accumulo {
+geotrellis.accumulo {
   catalog = "metadata"
   threads {
     collection.read = default
